Finished with the js may add enter capture, submit display
Finished up the basic concept. For final touches I need to add in the
text that goes with the input, capture the enter so that it won't submit
but instead goto next field (or submit on last), and then add in a
show/hide for the submit button that only shows on completion.